requirements.txt
kazoo==2.8.0
pyyaml==5.3.1
requests==2.24.0
psutil==5.7.2
protobuf==3.12.4
graphviz==0.14.1
pyzmq==19.0.2
pyqt5==5.15.0